RobotBuilder Sample Project
This sample project was created for the August 2015 WRRF workshop to
demonstrate how to use the RobotBuilder software to generate the
framework of a roboRIO robot project. It used the Team 100 roboRIO test
board and a LEGO worm drive arm with hi/lo arm position limit switches
and a potentiometer measuring arm angle. The arm was driven with a
continuous rotation servo motor.